+++ core/src/main/java/com/adatao/ddf/ml/ISupportML.java
@@ -55,5 +55,5 @@ public interface ISupportML extends IHandleDDFFunctionalGroup {
   /*
    * Compute the binary confusion matrix of this DDF based on the given model.
   */
-  Double[][] getConfusionMatrix(IModel model, double threshold) throws DDFException;
+  Long[][] getConfusionMatrix(IModel model, double threshold) throws DDFException;
 }
